+++ Sources/SwiftUIKit/Lists/ListSectionTitle.swift
@@ -10,22 +10,47 @@ import SwiftUI
 
 /**
  This view mimics the `Section` title in a grouped `List`.
+ 
+ The view doesn't add any insets by default, but you can use
+ the `withInsets` init parameter to make it apply a standard
+ grouped title padding to the view.
  */
 public struct ListSectionTitle: View {
 
     /**
      Create a list section title.
      */
-    public init(_ text: String) {
+    public init(
+        _ text: String,
+        withInsets: Bool = false
+    ) {
         self.text = text
+        self.applyInsets = withInsets
     }
 
     private let text: String
+    private let applyInsets: Bool
     
     public var body: some View {
         Text(text.uppercased())
             .foregroundColor(.secondary)
             .font(.footnote)
+            .withGroupedListSectionHeaderInsets(if: applyInsets)
+    }
+}
+
+private extension View {
+    
+    @ViewBuilder
+    func withGroupedListSectionHeaderInsets(
+        if condition: Bool
+    ) -> some View {
+        if condition {
+            self.padding(.leading)
+                .padding(.top, -3)
+        } else {
+            self
+        }
     }
 }